Until we have a tokenized field for resource path we need to do filename searches in the id field. Solr does not accept query beginning with wildcard, so we must know the prefix. Prefix is @base if not using --nobase in index (in which case prefix is "/").
We will probably have a --prefix parameter in the hook script for those who want to index multiple hosts or parent paths. With the current search in id field, the same prefix must be set in the reposSearchIdPrefix variable in javascript. With a tokenized copyField for the id that would probably not be needed at all.

git-svn-id: https://labs.repos.se/data/search/trunk@599 5a4de678-0f31-0410-84e4-bbaf904ce06c