Reworked folder copy options handling. The default behavior is to index contents of folder copies as individual add of folders and files inside target. For files change inside a copy in the same commit, this means one indexing of Add followed by Update (assuming that the change list always returns the parent folder before the sub-item).
However, also by default, indexing is skipped for "branch operations", to avoid huge indexes with multiple hits for everything in the typical software version control scenario. Instead contents of copies are only indexed if updated, meaning that changed files inside a branch will be searchable.

Behavior can be changed using the "foldercopy" command line option.

For indexing implemented using the change handler concept, i.e. everything but svnhead, additional flexibility is allowed. The method isHandleFolderCopyAsRecursiveAdd can be overridden to decide case by case if a copy should be indexed. This is however only called if the hook handles the folder copy. so set foldercopy=yes to get this call for all folder copies.

The test_docs unit test uses a post-commit hook with the same settings for every revision, so it only tests the "nobranch" mode. I've tested the rest manually. This architecture is unfortunate because change detection and indexing is intertwined, a design that should be avoided for 2.0.


git-svn-id: https://labs.repos.se/data/search/trunk@1058 5a4de678-0f31-0410-84e4-bbaf904ce06c