Added authorization to search results so that only users authorized to view files will see those files returned in their search results.
authz_parser.py: This Python module parses a Subversion `authz` file and
provides methods to authorize a user to view a specific path.

index.py: This Python CGI script replaces index.php and duplicates its
functionality, but also adds a function to authorize a user with authz_parser.py
for each search result before returning the results to the user.

repos-search.load.js: This JavaScript script was modified to (A) avoid caching
the results of the search query so that authorization changes will be reflected
immediately and (B) display a message indicating how many of the total results
the user is authorized to view.